backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
psycopg[binary]==3.2.13
gunicorn==23.0.0
python-dotenv
Flask-Cors
werkzeug
PyJWT
